requirements.txt
pytest==7.4.2
requests==2.31.0
PyYAML==6.0
jsonschema==4.19.0
pytest-html==3.2.0
pytest-cov==4.1.0
